+++ interface_stone_smart_flutter/lib/utils/extensions/payment_type_call_ext.dart
@@ -25,6 +25,8 @@ extension PaymentTypeCallExt on PaymentTypeCall {
       PaymentTypeCall.printFromBase64: "paymentPrintBase64",
       PaymentTypeCall.printWrapPaper: "paymentPrintWrapPaper",
       PaymentTypeCall.activeDebugLog: "paymentActiveDebugLog",
+      PaymentTypeCall.getPosSerialNumber: "paymentGetSerialNumber",
+      PaymentTypeCall.getPosManufacture: "paymentGetManufacture",
     };
     return paymentTypeCallMap[this] ?? (throw 'Not Implemented');
   }