+++ stone_smart_flutter/android/src/main/java/com/qztech/stone_smart_flutter/payments/PaymentsPresenter.java
@@ -115,6 +115,14 @@ public class PaymentsPresenter {
         mUseCase.printWrapPaper(lines, context);
     }
 
+    public String getPosSerialNumber() {
+        return mUseCase.getSerialNumber();
+    }
+
+    public String getPosManufacture() {
+        return mUseCase.getPosManufacture();
+    }
+
     public void dispose() {
         if (mSubscribe != null) {
             mSubscribe.dispose();