+++ stone_smart_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.3.2
+
+- Add functions `getPosSerialNumber` and `getPosManufacture`
+
 ## 1.3.1+1
 
 - Use interface_stone_smart_flutter version 1.0.3